pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "superconduct-tc-reg"
version = "0.0.0"
description = ""
readme = "README.md"
authors = ["Yahor Paromau"]
packages = [{ include = "superconduct_tc_reg", from = "./src" }]

[tool.poetry.dependencies]
python = ">3.10,<3.13"
torch = { version = "^2.0", source = "torch-cu118" }
mlflow = "^2.15.0"
scikit-learn = "^1.6.0"
numpy = ">=2.0,<2.1"
pandas = "^2.2.3"
torchmetrics = "^1.4.0.post0"
torchinfo = "^1.8.0"
tqdm = "^4.64"
psutil = "^6.0.0"
lightning = "^2.3.3"
hydra-core = "^1.3.2"
hydra-optuna-sweeper = "^1.2.0"
hydra-joblib-launcher = "^1.2.0"
dvc = {extras = ["http"], version = "^3.55.0"}
async-timeout = "^4.0.3"
xgboost = "^2.1.4"
dagshub = "^0.5.9"
hydra-submitit-launcher = "^1.2.0"
onnx = "^1.17.0"
onnxruntime = "^1.21.0"
skl2onnx = "^1.18.0"
onnxmltools = "^1.13.0"

[[tool.poetry.source]]
name = "torch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.group.interactive]
optional = true

[tool.poetry.group.interactive.dependencies]
ipykernel = "~6.16.0"
seaborn = "^0.12.2"
matplotlib = "^3.7.1"
plotly = "^5.14.1"
papermill = "^2.4.0"
nbconvert = "^7.16.4"
